Accept new max_cpu_frequency in host config
This commit adds new MaxCPUFrequency to map the new max_cpu_frequency
parameter in the host configuration. This parameter can be operated
by the host API.

This commit also converts the indentation to spaces in a page to
prevent indentation problem.

Test Pass:
Test to reconcile this value by deploying an AIOSX node with the
max_cpu_frequency configured by the Deployment Manager
Signed-off-by: Yuxing Jiang <yuxing.jiang@windriver.com>