Adding VIM API for sw-patch strategy
This commit introduces a VIM API in order to let DM to create, apply,
abort and delete sw-patch strategies.

The idea of these changes is to add the structure of the VIM API. The
sw-patch url may be changed on later commits, depending on the
strategy needed to handle from DM.

This also adds unit tests for sw-patch API calls.

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>